Docs and rename postgres -> postgresql
