+++ mangum/adapter.py
@@ -58,8 +58,6 @@ class Mangum:
     log_level: str = "info"
     api_gateway_base_path: typing.Optional[str] = None
     text_mime_types: typing.Optional[typing.List[str]] = None
-    api_gateway_endpoint_url: typing.Optional[str] = None
-    api_gateway_region_name: typing.Optional[str] = None
     ws_config: typing.Optional[dict] = None
 
     def __post_init__(self) -> None:
@@ -174,11 +172,11 @@ class Mangum:
         stage = event["requestContext"]["stage"]
         domain_name = event["requestContext"]["domainName"]
 
-        api_gateway_endpoint_url = (
-            self.api_gateway_endpoint_url or f"https://{domain_name}/{stage}"
+        api_gateway_endpoint_url = self.ws_config.get(
+            "api_gateway_endpoint_url", f"https://{domain_name}/{stage}"
         )
-        api_gateway_region_name = (
-            self.api_gateway_region_name or os.environ["AWS_REGION"]
+        api_gateway_region_name = self.ws_config.get(
+            "api_gateway_region_name", os.environ["AWS_REGION"]
         )
 
         websocket = WebSocket(
